+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 - Move `@task_only` to `djangae.decorators`
 - Add `@task_or_superuser_only` and `@csrf_exempt_if_task`.
 - Allow `defer_iteration_with_finalize` to use different sharding strategy
+- Allow custom callable for `djangae.processing.datastore_key_ranges` to generate random keys
 
 ### Bug fixes: